refactor: 지도 컴포넌트 이름 변경 및 폴더 이동
