+++ app/page.tsx
@@ -1,4 +1,4 @@
-import Map from "./(map)/page";
+import Map from "./(map)/_components/Map";
 
 export default function Home() {
   return <Map />;
